sentiment-analyzer-backend/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3

# AI/ML dependencies - using more compatible versions
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0

# Data processing
pandas>=2.0.0
numpy>=1.21.0

# AWS dependencies
boto3>=1.26.0
botocore>=1.29.0

# Environment and configuration
python-dotenv>=0.19.0

# Utility dependencies
Werkzeug>=2.2.0
Jinja2>=3.1.0
itsdangerous>=2.1.0
click>=8.0.0
MarkupSafe>=2.1.0

# HTTP and networking
urllib3>=1.26.0
requests>=2.28.0
certifi>=2022.0.0
charset-normalizer>=3.0.0
idna>=3.0.0

# Date/time handling
python-dateutil>=2.8.0
pytz>=2022.0

# Development and debugging (optional)
gunicorn>=20.0.0  # For production deployment



